@@ -1,6 +1,6 @@
 {
   "name": "n8n-nodes-pocketbase",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "description": "Interact with your PocketBase instance",
   "keywords": [
     "n8n-community-node-package"
@@ -33,22 +33,23 @@
       "dist/credentials/PocketBaseApi.credentials.js"
     ],
     "nodes": [
-      "dist/nodes/PocketBase/PocketBase.node.js"
+      "dist/nodes/PocketBase/PocketBase.node.js",
+      "dist/nodes/PocketBaseSend/PocketBaseSend.node.js"
     ]
   },
   "devDependencies": {
-    "@types/express": "^4.17.6",
     "@types/request-promise-native": "~1.0.15",
-    "@typescript-eslint/parser": "~5.45",
-    "eslint-plugin-n8n-nodes-base": "^1.11.0",
-    "gulp": "^5.0.0",
-    "n8n-core": "*",
-    "n8n-workflow": "*",
-    "prettier": "^2.7.1",
-    "typescript": "~4.8.4"
+    "@typescript-eslint/parser": "^7.15.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-n8n-nodes-base": "^1.16.3",
+    "gulp": "^4.0.2",
+    "prettier": "^3.3.2",
+    "typescript": "^5.5.3"
+  },
+  "peerDependencies": {
+    "n8n-workflow": "*"
   },
   "dependencies": {
-    "cross-fetch": "^3.1.6",
-    "pocketbase": "^0.15.2"
+    "pocketbase": "^0.15.3"
   }
 }
